+++ SwiftClip/MenuBar/StandalonePopupMenuBuilder.swift
@@ -8,6 +8,6 @@ enum StandalonePopupMenuBuilder {
         HistoryMenuSection.addStandalone(to: menu, environment: environment, target: target)
         menu.addItem(.separator())
         SnippetsMenuSection.addStandalone(to: menu, environment: environment, target: target)
-        ActionMenuSection.add(to: menu, target: target)
+        ActionMenuSection.add(to: menu, target: target, showsQuitShortcut: false)
     }
 }